refactor(capabilities): make anthropic own its kinds
Closes #2173
