+++ crates/aether-kinds/src/descriptors.rs
@@ -51,14 +51,13 @@ mod tests {
     use aether_data::{Primitive, SchemaType};
 
     use crate::{
-        CliSend, CliSendResult, Delete, DeleteResult, DrawTriangle, DropComponent, DropResult,
-        FsFetch, FsFetchResult, Key, LifecycleUnsubscribeAll, List, ListResult, LoadComponent,
-        LoadResult, LyriaGenerate, LyriaGenerateResult, Mat4Apply, MessagesSend,
-        MessagesSendResult, MouseButton, MouseMove, NanobananaGenerate, NanobananaGenerateResult,
-        Ping, Pong, ProcessExited, Read, ReadResult, RecordResult, ReplaceComponent, ReplaceResult,
-        Spawn, SpawnResult, SubscribeInput, SubscribeInputResult, Terminate, TerminateResult, Tick,
-        TrajectoryEnd, TrajectoryLog, TrajectorySample, UnsubscribeAll, UnsubscribeInput, Write,
-        WriteResult,
+        Delete, DeleteResult, DrawTriangle, DropComponent, DropResult, FsFetch, FsFetchResult, Key,
+        LifecycleUnsubscribeAll, List, ListResult, LoadComponent, LoadResult, LyriaGenerate,
+        LyriaGenerateResult, Mat4Apply, MouseButton, MouseMove, NanobananaGenerate,
+        NanobananaGenerateResult, Ping, Pong, ProcessExited, Read, ReadResult, RecordResult,
+        ReplaceComponent, ReplaceResult, Spawn, SpawnResult, SubscribeInput, SubscribeInputResult,
+        Terminate, TerminateResult, Tick, TrajectoryEnd, TrajectoryLog, TrajectorySample,
+        UnsubscribeAll, UnsubscribeInput, Write, WriteResult,
     };
 
     #[test]
@@ -128,10 +127,6 @@ mod tests {
         assert!(names.contains(&Terminate::NAME));
         assert!(names.contains(&TerminateResult::NAME));
         assert!(names.contains(&ProcessExited::NAME));
-        assert!(names.contains(&MessagesSend::NAME));
-        assert!(names.contains(&MessagesSendResult::NAME));
-        assert!(names.contains(&CliSend::NAME));
-        assert!(names.contains(&CliSendResult::NAME));
         assert!(names.contains(&NanobananaGenerate::NAME));
         assert!(names.contains(&NanobananaGenerateResult::NAME));
         assert!(names.contains(&LyriaGenerate::NAME));
@@ -162,40 +157,6 @@ mod tests {
                 .iter()
                 .find(|d| d.name == name)
                 .expect("test setup: gemini result kind is registered in descriptor inventory");
-            assert!(
-                matches!(d.schema, SchemaType::Enum { .. }),
-                "{name} should be Enum, got {:?}",
-                d.schema
-            );
-        }
-    }
-
-    #[test]
-    fn anthropic_requests_are_structured_schemas() {
-        // ADR-0050: the two send kinds carry String/Vec/Option fields,
-        // so they must serialize as non-cast (structured) structs — the
-        // hub builds them from agent params via the wire encoder.
-        let descs = all();
-        for name in [MessagesSend::NAME, CliSend::NAME] {
-            let d = descs
-                .iter()
-                .find(|d| d.name == name)
-                .expect("test setup: anthropic request kind is registered in descriptor inventory");
-            let SchemaType::Struct { repr_c, .. } = &d.schema else {
-                panic!("{name} should be Struct, got {:?}", d.schema);
-            };
-            assert!(!*repr_c, "{name} contains String/Vec, must be structured");
-        }
-    }
-
-    #[test]
-    fn anthropic_results_are_enum_schemas() {
-        let descs = all();
-        for name in [MessagesSendResult::NAME, CliSendResult::NAME] {
-            let d = descs
-                .iter()
-                .find(|d| d.name == name)
-                .expect("test setup: anthropic result kind is registered in descriptor inventory");
             assert!(
                 matches!(d.schema, SchemaType::Enum { .. }),
                 "{name} should be Enum, got {:?}",
